Cleanup. Move FrameSender to its own file
